fix: Fixes race condition in ws.Subscribe
When the web socker service Subscribe function is called while an
already running goroutine receives an onEvent message with type
`subscribed`, the map `pubChanIDs` is accessed in both `Subscribe`
function, as a read operation, and `onEvent` function, as a write
operation.  This happens at the same time causing a race condition.

Fix involved places a mutex lock inside the `Subscribe` function.
The `onEvent` function already has a mutex lock.