Fix 100% cpu usage on websocket dial error
Bug:

When the websocket dialling fails (the server could be down for
example), the process enters in some read loop, and gets stuck at 100%
cpu.

I'm not sure why, but when we signal a close on this channel
(w.downstream) the process loops somewhere, and the CPU gets stuck at
100%.

We don't really need this, since the user is still trying to connect,
and is still not listening for any message, thus he is not expecting the
downstream channel to receive a close message.